pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ekwize <ulanovsega@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.25.0"
fastapi = {extras = ["all"], version = "^0.108.0"}
SQLAlchemy = "^2.0.25"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.6"
flower = "^2.0.1"
pillow = "^10.2.0"
sqladmin = "^0.16.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
isort = "^5.13.2"
autoflake = "^2.2.1"
alembic = "^1.13.1"
python-json-logger = "^2.0.7"
sentry-sdk = {extras = ["fastapi"], version = "^1.39.2"}
fastapi-versioning = "^0.10.0"
gunicorn = "^21.2.0"
prometheus-fastapi-instrumentator = "^6.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "src"]
